Make bin/e compatible with Linux.
The revision deals with POSIX standards for shells. The old version uses
the bashism "test" which will cause an error on some other shells.
Specifically, in Ubuntu 8+ (and probably most Debian versions) /bin/sh
is linked to /bin/dash which more stricly follows the POSIX standard. So
now it follows the standard and is more likely to work on most systems.