zsh: switch back to emacs keybindings
Vim keybindings in the shell work okay, but the lack of a status line
slowed me down quite a bit. And the lack of ctrl-r for history searching
wasn't great. Maybe I'll go back to Vim keybindings later once I figure
out a good way to solve these issues, but for now it's back to the
default emacs keybindings.