tmux: fix cross-platform copy
Fixing copy-paste in tmux on OSX will now no longer interfere with tmux
on other OS's. Previously, the tmux config included
``reattach-to-user-namespace``, which is a hacky-script to fix tmux on
OSX. But since the tmux config is limited, you cannot specify to only
run the script on OSX and so tmux will crash if you attempt to launch it
on other systems where the script does not exist.