vim python: properly escape file paths
I like to run Python code with ``<leader>r``, which is mapped to
``python %:p``, where ``%:p`` should give the absolute path to the
current file. Unfortunately, ``%:p`` doesn't shell escape spaces in the
file path and so can cause issues (e.g. when running on a file inside
"~/Google Drive"). With this commit, the file path will be shell escaped
and therefore can work in file paths that contain spaces.