vim: use TeXShop as PDF view for LaTeX
TeXShop as a PDF viewer has the benefit of being able to deal with
synctex requests (which allow the PDF to be updated, without losing your
current location). Preview.app doesn't have this feature, but other PDF
viewers do (e.g. Skim). For now, I'm going to try TeXShop, but this may
change.