latexmk: switch to lualatex
I had been using xelatex as the presentation theme I was using (mtheme)
required it for the theme fonts. Before that I just used pdflatex (which
worked fine). But I ran into an issue recently where a journal article
style (for ASME) was not compatible with xelatex (the PDF would be
generated, but it messed up page margins, etc.). So my option was either
manually run pdflatex for the ASME style (which breaks my workflow) or
xelatex for the presentations with pdflatex as the default (which also
breaks my workflow). Fortunately, it seems like lualatex works with both
the ASME style and the presentation files, without any need for futher
tweaking. So I think I'll switch to lualatex for now.