vim: ale: minimize linter lag for python
The ale vim plugin (for running linters asynchronously) was causing huge
CPU spikes and slowing down VIM so much that it became practically
unusable. This only starting happening recently, so must have been
related to a change in the ale codebase. Regardless, it seems explicitly
selecting flake8 as the python linter (instead of the default of
multiple linters) prevents the issue.