git: verbose git commits by default
Switch the git commit aliases (gc and gca) to default to run with the
verbose flag, which shows a diff of each file included in the commit.
I've found this very helpful in my workflow as of late and therefore
want to save time by changing the default to use the verbose flag
(rather than typing ``gc -c <file>`` or similar everytime).