Add conda config (for Anaconda)
Add a conda config file. The big reason is to disable the auto-activate
base env default for the new conda init functionality, which was causing
issues with tmux. Specifically, it seems with the current version of
conda init zsh (which activates the base conda env), tmux sessions don't
properly respect conda env paths, thereby making it where a conda env
can be activate inside tmux, but the system-level python still gets
called. The apparent fix to this issue is to deactivate any conda envs
before opening a new tmux session, then activating the conda env inside
the tmux session. It doesn't make total sense, but it does fix the issue
(for now).