hydra: remove unused commands
I previously added commands for splitting windows to take up the top
half or bottom half of the screens, which could be useful for a second
display in portrait (vertical) orientation. However, I haven't found
portrait mode as useful as I thought and the shortcut keys I selected
for the two modes interferred with some existing shortcuts. So removing
the code makes sense.