nvim: revert `Y` behavior to yank entire line
Revert the behavior of `Y` to yank the entire line, regardless of the
cursor position. With nvim v0.6, the default for `Y` switched to yank
from the cursor position to the end of line, with `yy` being the way to
yank the entire line (similar to how `dd` is delete the entire line).
Keeping `Y` as yank the entire line is basically a personal preference
point for me, mostly due to muscle memory of `Y`.