nvim: set termguicolors before colorscheme
Not entirely sure why, but nvim starting throwing an error with my
Solarized colorscheme (solarized8.vim) that implied `termguicolors` was
not being set properly. This may be related to my upgrade to the newest
macOS version, but regardless, the fix seems to be to enable
`termguicolors` before setting the colorscheme. (Whereas before I was
setting the colorscheme first, then `termguicolors` in my nvim
`init.vim` file).