nvim: only show errors on current selection
The Async Linter Engine (ALE) switched a default setting where it showed
all errors and warnings in the file via virtual text (text displayed on
top of the code but not saved in the file). This was too overwhelming
for my tastes and so I've switched the setting to only show the problems
for the current line.